requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
Pillow==8.3.2
numpy==1.21.2 